CORS settings fix

+++ db/migrations/000097_business_accounts.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS business_accounts (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
-    business_user_id BIGINT NOT NULL,
+    id INT AUTO_INCREMENT PRIMARY KEY,
+    business_user_id INT NOT NULL,
     seats_total INT NOT NULL DEFAULT 0,
     seats_used INT NOT NULL DEFAULT 0,
     status ENUM('active', 'suspended') NOT NULL DEFAULT 'active',
@@ -11,11 +11,11 @@ CREATE TABLE IF NOT EXISTS business_accounts (
 );
 
 CREATE TABLE IF NOT EXISTS business_workers (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
-    business_user_id BIGINT NOT NULL,
-    worker_user_id BIGINT NOT NULL,
+    id INT AUTO_INCREMENT PRIMARY KEY,
+    business_user_id INT NOT NULL,
+    worker_user_id INT NOT NULL,
     login VARCHAR(255) NOT NULL,
-    chat_id BIGINT NOT NULL,
+    chat_id INT NOT NULL,
     status ENUM('active', 'disabled') NOT NULL DEFAULT 'active',
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP NULL DEFAULT NULL ON UPDATE CURRENT_TIMESTAMP,
@@ -29,11 +29,11 @@ CREATE TABLE IF NOT EXISTS business_workers (
 );
 
 CREATE TABLE IF NOT EXISTS business_worker_listings (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
-    business_user_id BIGINT NOT NULL,
-    worker_user_id BIGINT NOT NULL,
+    id INT AUTO_INCREMENT PRIMARY KEY,
+    business_user_id INT NOT NULL,
+    worker_user_id INT NOT NULL,
     listing_type ENUM('ad', 'service', 'work', 'work_ad', 'rent', 'rent_ad') NOT NULL,
-    listing_id BIGINT NOT NULL,
+    listing_id INT NOT NULL,
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     UNIQUE KEY uniq_worker_listing (listing_type, listing_id),
     INDEX idx_business_worker_listings_business (business_user_id),
@@ -43,8 +43,8 @@ CREATE TABLE IF NOT EXISTS business_worker_listings (
 );
 
 CREATE TABLE IF NOT EXISTS business_seat_purchases (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
-    business_user_id BIGINT NOT NULL,
+    id INT AUTO_INCREMENT PRIMARY KEY,
+    business_user_id INT NOT NULL,
     seats INT NOT NULL,
     amount DECIMAL(10, 2) NOT NULL DEFAULT 0.00,
     provider VARCHAR(255) DEFAULT NULL,
